backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.3.0
autobahn==23.6.2
Automat==25.4.16
billiard==4.2.2
bitarray==3.7.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
ckzg==2.1.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
cytoolz==1.0.1
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangochannelsrestframework==1.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.7.0
hexbytes==1.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.2
multidict==6.6.4
mysqlclient==2.2.7
packaging==25.0
parsimonious==0.10.0
prompt_toolkit==3.0.52
propcache==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pyunormalize==17.0.0
pywin32==311;platform_system == "Windows"
redis==6.4.0
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
Twisted==25.5.0
txaio==25.9.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
zope.interface==8.1.1
